refactor: use app.interface.ts instead of declarations.d.ts
