+++ generators/authentication/templates/ts/authentication.ts
@@ -1,15 +1,8 @@
-import { ServiceAddons } from '@feathersjs/feathers';
 import { AuthenticationService, JWTStrategy } from '@feathersjs/authentication';
 import { LocalStrategy } from '@feathersjs/authentication-local';
 import { expressOauth } from '@feathersjs/authentication-oauth';
 
-import { Application } from './declarations';
-
-declare module './declarations' {
-  interface ServiceTypes {
-    'authentication': AuthenticationService & ServiceAddons<any>;
-  }
-}
+import { Application } from './app.interface';
 
 export default function(app: Application) {
   const authentication = new AuthenticationService(app);
