+++ generators/connection/templates/ts/knex.ts
@@ -1,5 +1,5 @@
 import knex from 'knex';
-import { Application } from './declarations';
+import { Application } from './app.interface';
 
 export default function (app: Application) {
   const { client, connection } = app.get('<%= database %>');
