+++ generators/connection/templates/ts/sequelize.ts
@@ -1,5 +1,5 @@
 import { Sequelize } from 'sequelize';
-import { Application } from './declarations';
+import { Application } from './app.interface';
 
 export default function (app: Application) {
   const connectionString = app.get('<%= database %>');
