+++ generators/service/templates/ts/model/cassandra.ts
@@ -1,6 +1,6 @@
 // See https://express-cassandra.readthedocs.io/en/latest/schema/
 // for more of what you can do here.
-import { Application } from '../declarations';
+import { Application } from '../app.interface';
 
 export default function (app: Application) {
   const models = app.get('models');
