+++ generators/service/templates/ts/model/knex.ts
@@ -3,7 +3,7 @@
 // See http://knexjs.org/
 // for more of what you can do here.
 import Knex from 'knex';
-import { Application } from '../declarations';
+import { Application } from '../app.interface';
 
 export default function (app: Application) {
   const db: Knex = app.get('knexClient');
