+++ generators/service/templates/ts/model/sequelize.ts
@@ -1,7 +1,7 @@
 // See http://docs.sequelizejs.com/en/latest/docs/models-definition/
 // for more of what you can do here.
 import { Sequelize, DataTypes } from 'sequelize';
-import { Application } from '../declarations';
+import { Application } from '../app.interface';
 
 export default function (app: Application) {
   const sequelizeClient: Sequelize = app.get('sequelizeClient');
