+++ generators/service/templates/ts/types/memory.ts
@@ -1,5 +1,5 @@
 import { Service, MemoryServiceOptions } from 'feathers-memory';
-import { Application } from '<%= relativeRoot %>declarations';
+import { Application } from '<%= relativeRoot %>app.interface';
 
 export class <%= className %> extends Service {
   constructor(options: Partial<MemoryServiceOptions>, app: Application) {
