+++ generators/service/templates/ts/types/mongoose.ts
@@ -1,5 +1,5 @@
 import { Service, MongooseServiceOptions } from 'feathers-mongoose';
-import { Application } from '<%= relativeRoot %>declarations';
+import { Application } from '<%= relativeRoot %>app.interface';
 
 export class <%= className %> extends Service {
   constructor(options: Partial<MongooseServiceOptions>, app: Application) {
