+++ generators/service/templates/ts/types/nedb.ts
@@ -1,5 +1,5 @@
 import { Service, NedbServiceOptions } from 'feathers-nedb';
-import { Application } from '<%= relativeRoot %>declarations';
+import { Application } from '<%= relativeRoot %>app.interface';
 
 export class <%= className %> extends Service {
   constructor(options: Partial<NedbServiceOptions>, app: Application) {
