+++ generators/upgrade/templates/ts/authentication.ts
@@ -1,4 +1,4 @@
-import { Application } from '../../declarations';
+import { Application } from '../../app.interface';
 import { AuthenticationService, JWTStrategy } from '@feathersjs/authentication';
 import { LocalStrategy } from '@feathersjs/authentication-local';
 import { expressOauth } from '@feathersjs/authentication-oauth';
